LayoutTests/platform/mac-gpup/scrollingcoordinator/mac/latching/zero-delta-began-should-not-latch-expected.txt

FAIL: Timed out waiting for notifyDone to be called

Tests a gesture with zero delta on the began event doesn't latch on the document

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS overflow.scrollTop is 0
PASS window.pageYOffset is 0
